Use autocfg instead of a feature for try_exists
Per comments on https://github.com/andrewhickman/fs-err/pull/48 this moves from gating `try_exists` behind a feature to using `autocfg`. This is a build time library that inspects the current runtime and adds flags to the build so if a user is building with Rust 1.63 or higher then this feature will be enabled.