Implement Path::try_exists (#48)
* Implement Path::try_exists

Implements the `Path::try_exists` associated function mentioned in #46. The aliased function was stabilized in 1.63 so to preserve compatibility with older Rust versions (ci targets 1.40) a new feature flag was added.

* Use autocfg instead of a feature for try_exists

Per comments on https://github.com/andrewhickman/fs-err/pull/48 this moves from gating `try_exists` behind a feature to using `autocfg`. This is a build time library that inspects the current runtime and adds flags to the build so if a user is building with Rust 1.63 or higher then this feature will be enabled.

---------

Co-authored-by: Richard Schneeman <rschneeman@rschnee-ltmgmqp.internal.salesforce.com>
Co-authored-by: Andrew Hickman <andrew.hickman1@sky.com>