fs-err: Use AsRef<Path> to match std::fs
In order for fs-err to be a drop-in replacement for std::fs, it
needs to match the std::fs signature exactly.

Note that this can result in an extra allocation when the path provided
is already a PathBuf, but keeping the same API as std::fs is better than
over-optimizing for that case.

Fixes #34.